Re-rename pure_unlimited to pure
