+++ Cargo.toml
@@ -1,10 +1,10 @@
 [workspace]
 members = ["pyo3-stub-gen", "pyo3-stub-gen-derive"]
 exclude = [
-  # Note: "pure_abi3" and "pure_unlimited" cannot be built simultaneously
+  # Note: "pure_abi3" and "pure" cannot be built simultaneously
   #       since the ABI3 version of PyO3 is not compatible with the native version
   "examples/pure_abi3",
-  "examples/pure_unlimited",
+  "examples/pure",
   "examples/mixed",
   "examples/mixed_sub",
   "examples/mixed_sub_multiple",
