+++ examples/pure/Taskfile.yml
