+++ examples/pure/pure.pyi
