+++ examples/pure/src/bin/stub_gen.rs
@@ -2,7 +2,7 @@ use pyo3_stub_gen::Result;
 
 fn main() -> Result<()> {
     env_logger::Builder::from_env(env_logger::Env::default().filter_or("RUST_LOG", "info")).init();
-    let stub = pure_unlimited::stub_info()?;
+    let stub = pure::stub_info()?;
     stub.generate()?;
     Ok(())
 }
