+++ examples/pure/tests/test_python.py
@@ -1,4 +1,4 @@
-from pure_unlimited import sum, create_dict, read_dict, echo_path, ahash_dict
+from pure import sum, create_dict, read_dict, echo_path, ahash_dict
 import pytest
 import pathlib
